# openapi.yaml
openapi: 3.1.0
info:
  title: Eka Legal Consultancy API
  description: |
    Backend API for Eka Legal Consultancy practice management system.
    Supports client portal, booking system, case management, messaging, and document handling.
  version: 1.0.0
  contact:
    name: Eka Legal Support
    email: support@eka-legal.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.eka-legal.com/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "user-123"
        email:
          type: string
          format: email
          example: "john.doe@email.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [client, lawyer, admin]
        phone:
          type: string
          example: "+1 (555) 123-4567"
        avatarUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    LoginCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
      required:
        - email
        - password
    RegisterData:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        phone:
          type: string
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
        message:
          type: string

    # ============================================
    # Lawyer Profile & Public Content
    # ============================================
...
    ConsultationType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        price:
          type: number
          format: float
          description: Price in dollars (0 for free consultations)
        description:
          type: string
      required:
        - id
        - name
        - duration
        - price
        - description
    TimeSlot:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          example: "10:00"
        available:
          type: boolean
      required:
        - id
        - date
        - time
        - available
    Booking:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
          format: email
        consultationType:
          $ref: '#/components/schemas/ConsultationType'
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - clientEmail
        - consultationType
        - date
        - time
        - status
        - reason
        - createdAt
    CreateBookingRequest:
      type: object
      properties:
        consultationTypeId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        reason:
          type: string
      required:
        - consultationTypeId
        - date
        - time
        - name
        - email
        - reason

    # ============================================
    # Cases & Documents
    # ============================================
    Document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          description: MIME type
        size:
          type: integer
          description: File size in bytes
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
        url:
          type: string
          format: uri
      required:
        - id
        - name
        - type
        - size
        - uploadedAt
        - uploadedBy
        - url
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [note, document, status, meeting]
      required:
        - id
        - date
        - title
        - description
        - type
    Case:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, active, closed]
        caseType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
      required:
        - id
        - clientId
        - title
        - description
        - status
        - caseType
        - createdAt
        - updatedAt
        - documents
        - timeline
...
    DashboardStats:
      type: object
      properties:
        totalClients:
          type: integer
        activeCase:
          type: integer
        upcomingAppointments:
          type: integer
        pendingDocuments:
          type: integer
        appointmentsThisWeek:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              count:
                type: integer
      required:
        - totalClients
        - activeCase
        - upcomingAppointments
        - pendingDocuments
        - appointmentsThisWeek

    # ============================================
    # Intake Form
    # ============================================
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new client account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterData'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate user session
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the currently authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
                  format: password
                  minLength: 8
              required:
                - token
                - password
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid or expired token

  # ============================================
  # Public Endpoints
  # ============================================
...
  /booking/available-slots:
    get:
      tags:
        - Booking
      summary: Get available time slots
      description: Retrieve available booking slots for a specific date
      operationId: getAvailableSlots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date to check availability (YYYY-MM-DD)
      responses:
        '200':
          description: Available slots retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TimeSlot'
  /booking/bookings:
    get:
      tags:
        - Booking
      summary: Get my bookings
      description: Retrieve all bookings for the authenticated user
      operationId: getMyBookings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Booking'
        '401':
          description: Not authenticated

    post:
      tags:
        - Booking
      summary: Create booking
      description: Create a new consultation booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid booking request
  /booking/bookings/{bookingId}:
    delete:
      tags:
        - Booking
      summary: Cancel booking
      description: Cancel an existing booking
      operationId: cancelBooking
      security:
        - BearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Booking not found

  # ============================================
  # Case Endpoints
  # ============================================
  /cases:
    get:
      tags:
        - Cases
      summary: Get my cases
      description: Retrieve all cases for the authenticated user
      operationId: getMyCases
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, active, closed]
          description: Filter cases by status
      responses:
        '200':
          description: Cases retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Case'
        '401':
          description: Not authenticated
  /cases/{caseId}:
    get:
      tags:
        - Cases
      summary: Get case by ID
      description: Retrieve a specific case by its ID
      operationId: getCaseById
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Case retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Case'
        '404':
          description: Case not found

  # ============================================
  # Document Endpoints
  # ============================================
...
  /dashboard/client/stats:
    get:
      tags:
        - Dashboard
      summary: Get client dashboard stats
      description: Retrieve dashboard statistics for clients
      operationId: getClientStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Client stats retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardStats'
  /dashboard/lawyer/stats:
    get:
      tags:
        - Dashboard
      summary: Get lawyer dashboard stats
      description: Retrieve dashboard statistics for lawyers
      operationId: getLawyerStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lawyer stats retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardStats'

  # ============================================
  # Intake Form Endpoints
  # ============================================
...